+++ packages/pwa-kit-dev/src/configs/webpack/loaders/extensions-loader.test.js
@@ -5,9 +5,11 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import ExtensionLoader from './extensions-loader'
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 import dedent from 'dedent'
+import compiler from './test/compiler'
+
+import ExtensionLoader from './extensions-loader'
 
 const mockConfig = {
     app: {
@@ -20,37 +22,55 @@ const mockConfig = {
     }
 }
 
+const mockPkgConfig = {
+    devDependencies: {
+        '@salesforce/extension-this': '1.0.0-dev',
+        '@salesforce/extension-that': '1.0.0-dev',
+        '@companyx/extension-another': '1.0.0-dev',
+        'extension-this': '1.0.0-dev'
+    }
+}
+
 jest.mock('@salesforce/pwa-kit-runtime/utils/ssr-config', () => {
     return {
         getConfig: jest.fn(() => mockConfig)
     }
 })
 
-describe('Extenion Loader', () => {
+describe('Extension Loader', () => {
     beforeAll(() => {
         getConfig.mockImplementation(() => mockConfig)
     })
 
-    test('Returns single file re-exporting all extensions configured.', () => {
-        expect(ExtensionLoader()).toBe(dedent`
+    test('Returns single file re-exporting all extensions configured.', async () => {
+        const stats = await compiler('mock-extensions.js', {pkg: mockPkgConfig})
+        const output = stats.toJson({source: true}).modules[1].source
+
+        expect(output).toBe(dedent`
             /*
             * Copyright (c) 2024, salesforce.com, inc.
             * All rights reserved.
             * SPDX-License-Identifier: BSD-3-Clause
             * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
             */
-            // Extension Imports
+            import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
+
+            // App Extensions
             import SalesforceThis from '@salesforce/extension-this/setup-app'
             import SalesforceThat from '@salesforce/extension-that/setup-app'
             import CompanyxAnother from '@companyx/extension-another/setup-app'
             import This from 'extension-this/setup-app'
 
-            export default [
-                new SalesforceThis({}),
-                new SalesforceThat({}),
-                new CompanyxAnother({}),
-                new This({})
-            ]
+            export const getExtensions = () => {
+                const configuredExtensions = getConfig()?.app?.extensions || []
+
+                return [
+                    configuredExtensions.includes('@salesforce/extension-this') ? new SalesforceThis({}) : false,
+                    configuredExtensions.includes('@salesforce/extension-that') ? new SalesforceThat({}) : false,
+                    configuredExtensions.includes('@companyx/extension-another') ? new CompanyxAnother({}) : false,
+                    configuredExtensions.includes('extension-this') ? new This({}) : false
+                ].filter(Boolean)
+            }
         `)
     })
 })