+++ packages/pwa-kit-dev/src/configs/webpack/loaders/test/mock-extensions.js
@@ -5,10 +5,8 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import ApplicationExtension from './application-extension'
-
 // This file exports the default extensions object, which is empty. During local development and the production
 // build process, this object is replaced with those application extensions configured by the base PWA-Kit project.
-const extensions: ApplicationExtension[] = []
+const extensions = []
 
 export default extensions