+++ packages/pwa-kit-react-sdk/src/ssr/universal/extensibility/index.ts
@@ -9,4 +9,5 @@
 // is broken.
 
 export {default as extensions} from './extensions'
+export {default as ApplicationExtension} from './application-extension'
 export * from './types'