+++ packages/pwa-kit-react-sdk/src/ssr/universal/extensibility/utils.ts
@@ -7,7 +7,7 @@
 
 import React from 'react'
 import {applyHOCs} from '../utils'
-import {IApplicationExtension} from './types'
+import ApplicationExtension from './application-extension'
 
 /**
  * Given the provided Application, apply all the App extensions to it.
@@ -16,9 +16,14 @@ import {IApplicationExtension} from './types'
  */
 export const applyAppExtensions = (
     App: React.ComponentType,
-    extensions: IApplicationExtension[]
+    extensions: ApplicationExtension[]
 ): React.ComponentType => {
-    const extendAppHocs = extensions.map(({extendApp}) => extendApp).filter(Boolean)
+    const extendAppHocs = extensions
+        // TODO: All Application Extensions configuration objects will extend from a single IApplicationExtensionConfig type so that we
+        // can add a feature to toggle if the extension is enabled or disabled. This will happen in the tupal config support ticket.
+        // .map((extension) => extension.getConfig().enabled && extension.extendApp)
+        .map((extension) => extension.extendApp.bind(extension))
+        .filter(Boolean)
 
     return applyHOCs(App, extendAppHocs)
 }